requirements.txt
# == Existing pinned packages (from your snippet) ==
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
distro==1.9.0
dnspython==2.7.0
faiss-cpu==1.10.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
MarkupSafe==3.0.2
numpy==2.2.3
openai==1.63.2
packaging==24.2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11.1
python-dotenv==1.0.1
requests==2.32.3
rpy2==3.5.17
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
tzlocal==5.3
urllib3==2.3.0

# == Some data-science & ML packages referenced in your code ==
altair
beautifulsoup4
bokeh
catboost
cython
cvxpy
dash
dask
fastapi
fastparquet
flair
folium
gensim
geopandas
gradio
html5lib
ipywidgets
joblib
keras
lightgbm
lxml
matplotlib
networkx
nltk
numba
opencv-python
pandas
pillow
plotly
polars
prophet
pydantic  # (already pinned above, so normally skip or unify)
pyarrow
pyomo
pytest
pyyaml
scikit-image
scikit-learn
scipy
sentence-transformers
seaborn
spacy
sqlalchemy
statsforecast
statsmodels
streamlit
stanza
textblob
torch
transformers
tsfresh
tslearn
ujson
vaex
xgboost
